refactor: add aggregated modules and change project structure
