--- modules/createForm.mjs
@@ -1,5 +1,0 @@
-const createForm = (id, name, method, children) => {
-  return `<form method=${method} name=${name} id=${id}>${children.map((child) => child)}`;
-};
-
-export default createForm;
